Fleshing out module system slide deck
The start was re-structured to have a bigger focus on the fundamentals
of how the module system is implemented in the runtime, and then
provides more usage examples before going into how to write custom
modules.
This deck now also features submodules, and special types such as
listOf and attrsOf.